+++ roscopter_sim/include/roscopter_sim/gz_compat.h
